// API_Client/client/models/create-size-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { NameTranslationsDto } from './name-translations-dto';

export interface CreateSizeDto {
    /**
     * მრავალენოვანი სახელი — { ka: {name}, en?, ru? }, ka სავალდებულოა
     */
    'translations': NameTranslationsDto;
    /**
     * მოკლე კოდი (მაგ. \"2S\", \"3XL\")
     */
    'code': string;
}